refactor: move fetching of image dimensions to matrix client when uploading attachments
