+++ src/lib/chat/matrix/media.ts
@@ -20,6 +20,25 @@ export function readFileAsArrayBuffer(file: File | Blob): Promise<ArrayBuffer> {
   });
 }
 
+/**
+ * Get the dimensions of an image file.
+ * @param {File} file The image file to read.
+ * @return {Promise} A promise that resolves with an object containing
+ *   the width and height of the image when it is loaded.
+ */
+export function getImageDimensions(file: File): Promise<{ width: number; height: number }> {
+  return new Promise((resolve, reject) => {
+    const img = new Image();
+    img.onload = () => {
+      resolve({ width: img.width, height: img.height });
+    };
+    img.onerror = () => {
+      reject(new Error('Unable to determine image dimensions.'));
+    };
+    img.src = URL.createObjectURL(file);
+  });
+}
+
 export class UploadCanceledError extends Error {}
 
 // https://github.com/matrix-org/matrix-react-sdk/blob/develop/src/ContentMessages.ts#L339
