+++ src/store/messages/uploadable.ts
@@ -1,14 +1,12 @@
 import { CallEffect, call } from 'redux-saga/effects';
 
-import { FileType, getFileType, getImageDimensions } from './utils';
+import { FileType, getFileType } from './utils';
 import { Message } from '.';
 import { uploadFileMessage, uploadImageUrl } from '../../lib/chat';
 
-export const createUploadableFile = async (file): Promise<Uploadable> => {
+export const createUploadableFile = (file): Uploadable => {
   if (file.nativeFile && getFileType(file.nativeFile) === FileType.Media) {
-    const dimensions = await getImageDimensions(file.nativeFile);
-    console.log('XXXDIMENSIONS', dimensions);
-    return new UploadableMedia(file, dimensions.width, dimensions.height);
+    return new UploadableMedia(file);
   } else if (file.giphy) {
     return new UploadableGiphy(file);
   } else {
@@ -25,18 +23,14 @@ export interface Uploadable {
 export class UploadableMedia implements Uploadable {
   public optimisticMessage: Message;
 
-  constructor(public file, private width: number, private height: number) {}
+  constructor(public file) {}
   *upload(channelId, rootMessageId) {
-    console.log('xxxxdimensdwdwed', this.width, this.height);
-
     return yield call(
       uploadFileMessage,
       channelId,
       this.file.nativeFile,
       rootMessageId,
-      this.optimisticMessage?.id?.toString(),
-      this.width,
-      this.height
+      this.optimisticMessage?.id?.toString()
     );
   }
 }
